ci-builder: fix build (#10158)
Copies over `yq` and `just` into the final docker image. Older
versions of `ci-builder` are known to contain `just`, something
must have changed to have them no longer fully include it.
We need to tie the release of `ci-builder` with a bump of
the foundry version unless the foundry version we are using
is still available via nightly binary download. This is because
every time we build foundry in CI it produces a bad build
that results in an "illegal instruction" error.