op-bindings: delete `more` files (#10272)
* op-bindings: delete `more` files

The `op-bindings` more files existed to add
extra information to the bindings that were
helpful in particular for the L2 genesis generation
process. This commit deletes all more files as well
as the code for generating them. It also deletes
the top level `registry.go` file that used an `init`
function to populate public getters for accessing
the data in more files.

Now that the L2 genesis doesn't happen in Go, this
is all dead code. Removing this will greatly reduce
diff when smart contract code changes.

* cleanup: more dead code

* cleanup: decouple dead code