op-bindings: delete ast package (#10299)
This package canonicalized the ast ids that were placed
in the bindings `more` files because any change in the
smart contracts would cause a massive diff in the storage
layout for all storage layouts as they used the ast ids
to represent types. We needed to canonicalize them to have
deterministic bindings generation, otherwise CI would generate
the bindings in a non deterministic way. As we adopted new
solidity features, it would find edge cases in our ast canonicalize
algo, breaking CI. This code helped us for a long time but
now it is no longer used and can be retired.