contracts-bedrock: small fix to cleanup L2 genesis (#10326)
* contracts-bedrock: small fix to cleanup L2 genesis

Adds a `run()` function to the L2 genesis generation script
so that its possible to run the command without the `--sig`
argument. Also resolve some linting isssues, semgrep has
a rule that prefers `@notice` over `@dev` for natspec comments.
Also use a foundry cheatcode to set the chainid during execution
to ensure that the correct chainid is set to remove the need to
set it as a cli flag.

* devnet: remove extra flag to foundry script

* contracts-bedrock: fix test

* contracts-bedrock: update config

Turns on funding of genesis accounts

* contracts-bedrock: fix broken test

* contracts-bedrock: cleanup