contracts-bedrock: portal opaque data test (#10443)
* contracts-bedrock: add deposit event coverage

Adds deposit event differential coverage for the two
different places where the deposit tx event exists.
This ensures that a valid event ends up getting emitted.
Also modularizes the code for serializing the event so that
there are not footguns when computing the "opaque data".

* contracts-bedrock: revert diff to portal

Removes the diff to the portal, perhaps we want it
but test still passes without it

* contracts-bedrock: lint

* test: improve

* tests: update

* tests: fixup

* semver: fix

* lint: fix