op-node: cleanup readme (#10679)
* op-node: cleanup readme

The readme hasn't been updated in some time, this updates the readme
with the latest information as a lot of the information has been
deprecated. This will make it more intuitive for how to use the `op-node`
to generate the L2 genesis for a network.

Also adds a simple `doctoc` recipe to the `Makefile` as other `README`
files in the repo do the same thing.

* readme: update