fix: OptimismPortal fuzz flakes (#11818)
OptimismPortal fuzz tests were flaking because the base fee would
increase so much that there wouldn't be enough gas to burn for the
user to be able to buy the gas that they want. Introduces an
additional assumption that will prevent the elasticity multiplier
and max change denominator from being such that the maximum
change to the base fee won't cause us to run out of gas.