Use DA /put path from spec (#12081)
* Accept /put path as described in spec

In addition to the currently used `/put/` path to ease the migration.
See https://github.com/ethereum-optimism/optimism/issues/11499.

* alt-DA: write to /put path as described in spec

The spec mandates using `/put` and not `/put/`, so that is what we
should do.

Warning: This will break DA solutions that only accept `/put/` at the
moment. It is recommended that DA solutions support both paths for a
while, so that updating OP-stack can happen independently of the exact
DA implementation version.

Closes https://github.com/ethereum-optimism/optimism/issues/11499.