contracts: clean up deploy script abstractions (#12336)
* contracts: clean up deploy script abstractions

Move the implementation deployments into `deployImplementations`
and the initializations into the initizer function so that things
are grouped more logically. Even though we use OPSM for modern
deployments, the legacy code is still in the repo. This change
makes it easier to implement Standard L2 Genesis, where we depend
on the `OptimismPortal` being initialized before the `SystemConfig`,
since the `SystemConfig` makes calls to the portal during its
`initialize` call.

* op-e2e: remove dependency on contracts existing

Modularizes the functionality more

* lint: fix