op-service, op-deployer: Support marshaling systemconfig in pre-Holocene format (#12619)
op-deployer generates rollup configs, and those rollup configs are currently broken for older versions of op-node that don't support reading the EIP1559Params field in the SystemConfig. This PR adds a meta field within the SystemConfig that, when enabled, marshals the SystemConfig without the EIP1559Params field. This solution is backwards-compatible and minimally invasive, requiring no changes to the consensus-critical code that consumes the SystemConfig elsewhere.

Closes https://github.com/ethereum-optimism/optimism/issues/12615.