Makefile: Fix executable command detection (#12080)
`$(command -v geth)` is valid shell syntax, but it gets processed by
make first, so that this never gets correctly executed by the shell. As
a result, `make install-geth` and `make install-eth2-testnet-genesis`
are always executed and not, as intended, just when the commands are not
present.

The fix is easy: escape the dollar by duplicating it, so that make won't
do anything and the correct command reaches the shell.