op-service: Fix failing SSZ fuzz test (#13324)
The SSZ fuzz tests were not erroring when the `edOffset` was 508 and the `txOffset` was 514. This is because the transaction unmarshaler was reading the number of transactions in the payload as zero, but not checking to see if there was additional data left over. This should never be the case for a valid payload.

To verify, I ran the fuzz suite for 5 minutes:

```
fuzz: elapsed: 5m1s, execs: 12910962 (42267/sec), new interesting: 9 (total: 41)
```